+++ puma/src/main/java/com/dianping/puma/storage/bucket/WriteBucket.java
@@ -28,5 +28,5 @@ public interface WriteBucket extends LifeCycle {
 	 */
 	boolean hasRemainingForWrite();
 
-	long position();
+	int position();
 }
